server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
faiss-cpu==1.7.4
rank-bm25==0.2.2
sentence-transformers>=2.7.0
torch>=2.2.0
torchvision>=0.17.0
Pillow==10.1.0
scikit-learn==1.3.2
python-multipart==0.0.6
pydantic==2.5.0
pyarrow>=21.0.0
